% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_betas.R
\name{ml_chol}
\alias{ml_chol}
\title{Extract Cholesky decomposition matrix of correlated random variables of an mlogit model}
\usage{
ml_chol(mod)
}
\description{
Extract Cholesky decomposition matrix of correlated random variables of an mlogit model
}
