% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_betas.R
\name{ml_draw_betas}
\alias{ml_draw_betas}
\title{Draw random coefficients of individuals for an estimated mixed logit model}
\usage{
ml_draw_betas(mod, num.draws = 100, use.halton = FALSE, scale = NA)
}
\arguments{
\item{mod}{the estimated models}

\item{num.draws}{the number of individuals for which random coefficients shall be drawn}

\item{use.halton}{shall Halton sequences be used instead of psedudo-random numbers?}
}
\value{
A matrix with num.draws rows and one column for each explanatory variable that has random coefficients
}
\description{
Note that the function is still restricted and does not yet work for all sorts
of random parameter distributions that mlogit supports.
}
