% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{ml_tidy}
\alias{ml_tidy}
\title{Get tidy output of estimated coefficients of mlogit model}
\usage{
ml_tidy(mod, scale = NA)
}
\arguments{
\item{mod}{the estimated model}

\item{scale}{A numeric value or variable name used to scale coefficients and standard errors. Typical application would be a "price" variable if you estimated a product choice. Then utilities can be interpreted as willingness to pay.}
}
\description{
Get tidy output of estimated coefficients of mlogit model
}
