% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_changes.R
\name{show_P_changes}
\alias{show_P_changes}
\title{Helper function to show changes in choice probabilities / market shares}
\usage{
show_P_changes(org, new, digits = 1)
}
\description{
Helper function to show changes in choice probabilities / market shares
}
