% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{ml_predict}
\alias{ml_predict}
\title{Alternative predict function for mlogit objects}
\usage{
ml_predict(mod, newdata, J = NA, num.draws = 1000, alts = NA,
  use.halton = TRUE)
}
\description{
Alternative predict function for mlogit objects
}
